requirements.txt
alembic==1.16.4
altgraph==0.17.4
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
cssselect2==0.8.0
deep-translator==1.11.4
defusedxml==0.7.1
et_xmlfile==2.0.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
fonttools==4.59.0
fpdf2==2.8.3
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
macholib==1.16.3
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.0.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
requests==2.32.4
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.42
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.23.0
zopfli==0.2.3.post1
